pop messagre delete salle

+++ PlatformErpOdc/src/app/salle/salle.page.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2';
 import { DetailDesListesPage } from '../detail-des-listes/detail-des-listes.page';
 import { SalleServiceService } from '../services/salles/salle-service.service';
 import { UtilisateurService } from '../services/utilisateur/utilisateur.service';
@@ -34,6 +35,7 @@ export class SallePage implements OnInit {
      //!::::::::::::total perso ::::::::::::
 
     this.callSalle()
+
   }
 
   callSalle(){
@@ -51,13 +53,37 @@ export class SallePage implements OnInit {
 
   }
 
-  deleteSalleId(idSalle:number){
 
-    this.salleService.deleteSalle(this.Utilisateur.login,this.Utilisateur.password,idSalle).subscribe(data=>{
-      if(data.message=="ok"){
-        this.callSalle()
+
+
+  popDeleteSalle(idSalle:number){
+    Swal.fire({
+      position:'center',
+      title: 'Voulez-vous supprimer ?',
+      showDenyButton: true,
+      showCancelButton: false,
+      confirmButtonText: 'Oui',
+      denyButtonText: `Non`,
+      denyButtonColor:'red',
+      // cancelButtonText: 'Annuler',
+      cancelButtonColor:'red',
+      confirmButtonColor: 'green',
+      heightAuto: false,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        //Swal.fire('Saved!', '', 'success');
+        this.salleService.deleteSalle(this.Utilisateur.login,this.Utilisateur.password,idSalle).subscribe(data=>{
+          if(data.message=="ok"){
+            this.callSalle()
+          }
+        });
+       
+      } else if (result.isDenied) {
+        //Swal.fire('Changes are not saved', '', 'info');
+      //  this.route.navigate(['tirage'])
       }
     });
+ 
   }
-
 }
